Start near the middle of our space
By starting in the middle we avoid a rare case where we choose a
starting point in a corner and then happen to only pick points outside
our space, thus never finding a valid next point. This is extremely
rare: With the default 30 samples we have better than 99.98% odds of
picking at least 1 point within our space, and even then that's only if
we happen to pick a point very near a corner to start from.

Fixes #34